requirements.txt
python-dotenv~=1.0.0
selenium~=4.15.2
python-dateutil~=2.8.2